+++ hail/python/hail/experimental/function.py
@@ -3,9 +3,6 @@ from hail.ir import Apply, Ref, Renderer, register_session_function
 from hail.expr.types import HailType
 from hail.expr.expressions import construct_expr, anytype, expr_any, unify_all
 from hail.typecheck import typecheck, nullable
-import hail
-
-from decorator import decorator
 
 class Function(object):
     def __init__(self, f, param_types, ret_type, name):
@@ -16,24 +13,6 @@ class Function(object):
 
     def __call__(self, *args):
         return self._f(*args)
-
-
-@typecheck(param_types=hail.expr.types.hail_type)
-def udf(*param_types):
-
-    uid = Env.get_uid()
-
-    @decorator
-    def wrapper(__original_func, *args, **kwargs):
-        registry = hail.ir.ir._udf_registry
-        if uid in registry:
-            f = registry[uid]
-        else:
-            f = define_function(__original_func, *param_types, _name=uid)
-            registry[uid] = f
-        return f(*args, **kwargs)
-
-    return wrapper
 
 
 @typecheck(f=anytype, param_types=HailType, _name=nullable(str))
